requirements.txt
pyautogen>=0.4
chromadb>=0.5
langchain>=0.3
alpaca-trade-api>=3.2
polygon-api-client>=1.14
pandas>=2.2
numpy>=1.26
python-dotenv>=1.0
yfinance>=0.2
ta>=0.11
